Improve models package structure and move database.py outside
