+++ repost/models/__init__.py
@@ -1,4 +1,6 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, DateTime
-from sqlalchemy.orm import relationship
+from repost.database import Base
 
-from .database import Base
+from .comment import Comment
+from .post import Post
+from .resub import Resub
+from .user import User
