+++ repost/models/user.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Boolean, Column, Integer, String
 from sqlalchemy.orm import relationship
 
-from .database import Base
+from . import Base
 
 
 class User(Base):
